Reorganize tests to avoid pyarrow conflict
There is a conflict between the pyarrow arrow shared library
and the one built in this project. They cannot both be loaded into
the python interpreter at the same time due to the shared symbol names.

As a hacky workaround, this runs all pyarrow related test data
generation functions in a different interpreter.